refactor: preprocessor class for data refactor: mv AmplitudeModel to amp/amp.py feat: as_config for decay group refactor: pass decayconfig instead of decay struct to data mode
